Create similaritywav.py
It's an alternative method to compare wave file, the main difference is instead finding the peak frequency, this algorithm directly compare the frequency domain of each wave file with each other and compute the similarity.The function used is 'correlation' If similarity is over some threshold, we deduce as match.

It is kind of slow (it compares every chunk with each other) ,the running time grow exponentially.., but it does work sometimes too (I have only tested the smaller file, as the run time become infinity when comparing two large file). 

I wonder if this approach can assist in our original approach ?
The program is quite messy though, but basically it do the similar things like dividing into chunks and compare one chunk with another.